Move enum VerifierState to protocols layer
