+++ libvcx/src/api_lib/api_handle/proof.rs
@@ -149,7 +149,7 @@ pub mod tests {
 
     use crate::api_lib::api_handle::connection::tests::build_test_connection_inviter_requested;
     use crate::api_lib::api_handle::proof;
-    use crate::aries_vcx::handlers::proof_presentation::verifier::VerifierState;
+    use crate::aries_vcx::protocols::proof_presentation::verifier::state_machine::VerifierState;
 
     use super::*;
